requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
PyJWT==2.8.0
redis==5.0.1
Flask-WTF==1.2.1
python-dotenv==1.0.0
werkzeug==3.0.1